% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_quantile.R
\name{bin_quantile}
\alias{bin_quantile}
\title{Quantile binarization}
\usage{
bin_quantile(D, cols = 1:ncol(D), top = 0.75, group = 1, groupthreshold=1:group, reverse = FALSE, dropNA = TRUE)
}
\arguments{
\item{D}{A data frame}

\item{cols}{The indexes or names of the columns the users wants to binarize. Default = all the columns}

\item{top}{The probability of the quantile (in the interval [0,1]). For example, when top = 0.75, the bottom 75 percent of the values in the column will be tagged as FALSE, and the remaining 25 percent will be tagged as TRUE}

\item{group}{The interval of columns that should be processed together. Default = 1}

\item{groupthreshold}{What columns of each group to use to find the threshold. For example, when groupthreshold=3, the third column of each group will be used to determine the threshold of each group. Default = use all columns of a group}

\item{reverse}{A logical parameter; when reverse=TRUE, the bottom part of the value is considered TRUE, and the top is FALSE. Default = FALSE}

\item{dropNA}{A logical parameter: when dropNA=TRUE output rows containing at least one NA are deleted. Default = TRUE}
}
\value{
test$data: database with binarized listed columns

test$thresholds: thresholds used for the binarization (they follow the column order)
}
\description{
The function bin_quantile() allows to transform numeric columns of data frames in logical values based on a threshold. The threshold is computed from the distribution of the numeric values in the column(s), divided in quantiles as specified by the user.
}

