% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customCI.R
\name{customCI}
\alias{customCI}
\title{Custom CI}
\usage{
customCI(x, transactions, transNot, reuse = TRUE, alpha = 1.96)
}
\arguments{
\item{x}{A set of rules.}

\item{transactions}{The transaction data set used to mine the associations or a set of different transactions to calculate interest measures from (Note: you need to set reuse=FALSE in the later case).}

\item{transNot}{The negated transaction data set for calculations.}

\item{reuse}{A logical indicating if information in quality slot should be reuse for calculating the measures.}

\item{alpha}{The critical value from normal distribution}
}
\value{
output$lowCI: the lower bound of a confidence interval

output$highCI: the higher bound of a confidence interval.
}
\description{
This function creates a custom confidence intervals for rules based on different probabilities with the default of a 95 percent confidence interval.
}

