% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improvementcustomCI.R
\name{improvement_customCI_par}
\alias{improvement_customCI_par}
\title{Improvement Custom CI}
\usage{
improvement_customCI_par(rules, t = 0.1)
}
\arguments{
\item{rules}{An object of class rules. This object must include the column oddsRatio, added using the function myOddsRatio().}

\item{t}{The threshold of minimum improvement in odds ratio to declare a rule more important than all its subsets. Default = 0.1}
}
\value{
d: data frame containing one logical column "imp". Rules that do not show significant improvement are labeled as FALSE. This single column data frame can be added to the other quality measures of the object of class rules.
}
\description{
This function determines whether the OR of a rule is interesting or if its importance depends only on one of its subsets. The goal is to identify and possibly eliminate redundant rules with a Occam's Razor strategy
}

