% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonredundant.R
\name{nonredundant}
\alias{nonredundant}
\title{Nonredundant}
\usage{
nonredundant(rules, variables, lowCI, highCI)
}
\arguments{
\item{rules}{A data frame of rules. The data frame must have a column labeled rules with the variables used in each rule, and columns for both the lower and higher bounds of the odds ratio confidence interval. The columns for the the confidence interval should be labeled lowCIcustom and highCIcustom. The TestFrame from the list output of overfit_test() is already formated in this way.}

\item{variables}{This should be a vector of the names of each variable used in the making of the rules as they will appear in the rules. They should be in string form.}

\item{lowCI}{the name of the column with the lower bound of the confidence interval}

\item{highCI}{the name of the column with the higher bound of the confidence interval}
}
\value{
A logical vector of the length rules with non-redundant rules labeled TRUE and redundant rules labeled FALSE.
}
\description{
This function determines whether the OR of a rule is interesting or if it just a consequence of a parent rule. The goal is to eliminate redundant rules with a Occam's Razor strategy. 
It is very similar to improvementCI() but accepts a data frame input.It is assumed all rules in the data frame will be unique.
}

