% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddrules.R
\docType{package}
\name{oddrules-package}
\alias{oddrules}
\alias{oddrules-package}
\title{Oddrules}
\description{
Examines the odds ratios of rules created by package arules
}
\details{
For more help see the vignette or search for a function by name
}

