% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overfit_test.R
\name{overfit_test}
\alias{overfit_test}
\title{Overfit test}
\usage{
overfit_test(whole_training, testing, seed=1:3, trainingsize=0.5*nrow(whole_training), alpha = 1.598, support = 0.005)
}
\arguments{
\item{whole_training}{The training data set}

\item{testing}{The testing data set}

\item{seed}{A vector of values for randomly splitting the data set}

\item{trainingsize}{A vector of sizes for training sets to mine from}

\item{alpha}{A threshold for significance}

\item{support}{The minimum support for rule mining}
}
\value{
list: 3 matrices with the number of rules mined for each iteration of size and seed meeting minimum support, significance, and are nonredundant, and a dataframe of said successfull rules
}
\description{
Allows User to mine training set for rules several different times with different training sets and extract information.
}

