% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pollution.R
\docType{data}
\name{Pollution}
\alias{Pollution}
\title{Pollution}
\format{An object of class \code{data.frame} with 76795 rows and 31 columns.}
\usage{
Pollution
}
\description{
A data set containing pollution data and asthma attacks in the city of houston for different pollutants
}
\details{
Includes data for 6 different pollutants and up to 5 day lag for each pollutant and a column labeled "event" containing whether there were more hospitalized asthma attacks in children then average.

@format A data frame with 31 variables and 76795 observations
}
\keyword{datasets}

