% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_split.R
\name{random_split}
\alias{random_split}
\title{Random Split}
\usage{
random_split(sampleLog, seed=1, split=5000)
}
\arguments{
\item{sampleLog}{The data to be split into training and testing}

\item{seed}{The seed used for the random split. Default is 1.}

\item{split}{The number of rows to be put into the training set. The rest of the data will be put into the testing set.}
}
\value{
return$whole_training: data sorted into training

return$testing: data sorted into testing group
}
\description{
Splits data into testing and training sets
}

